// src/enums/System.ts
export enum SystemIdEnums {
    // 区域概览
    OVER_VIEW_ID = 1,
    // 碳排总览
    CARBON_OVER_VIEW_ID = 2,
    // 能耗在线
    ENERGY_ONLINE_ID = 3,
    // 碳足迹
    CARBON_FOOTPRINT_ID = 4,
    // 碳普惠
    CARBON_BENEFIT_ID = 5,
    // 项目能耗评估
    PROJECT_ENERGY_ASSESSMENT_ID = 6,
}